examples: fix unneeded type casting for monitoring values

+++ examples/src/monit_amp.c
@@ -149,10 +149,10 @@ int main (int argc, char *argv [])
             fprintf (stdout, "[client:monit_amp]: %lld.%.9ld "
                     "ch0/ch1/ch2/ch3 = %d %d %d %d\n",
                     (long long)ts.tv_sec, ts.tv_nsec,
-                    (int32_t) monit_data.amp_ch0,
-                    (int32_t) monit_data.amp_ch1,
-                    (int32_t) monit_data.amp_ch2,
-                    (int32_t) monit_data.amp_ch3);
+                    monit_data.amp_ch0,
+                    monit_data.amp_ch1,
+                    monit_data.amp_ch2,
+                    monit_data.amp_ch3);
         }
     }
